requirements.txt
Flask==1.1.2
gunicorn==20.1.0
numpy==1.19.5
torch==1.8.1
Pillow==7.2.0
opencv-python==4.2.0.34

